+++ src/routes/post/[...post]/+page.ts
@@ -1,16 +1,15 @@
 import { error } from "@sveltejs/kit";
+import { getPost } from "$lib/utils/posts";
 import type { PageData } from "./$types";
 
 export const load = (async ({ params }) => {
-  const post = await import(/* @vite-ignore */ `../../../../cec/${params.post}.md`);
-  const { default: page, metadata: data } = post;
-
-  if (!page) {
+  const post = getPost(params.post);
+  if (!post) {
     error(404, "not found");
   }
 
   return {
-    data,
-    content: page
+    data: post.metadata,
+    content: post.default
   };
 }) satisfies PageData;
